// src/catalog/generated.ts
// GENERATED by scripts/build-catalog.mjs - do not edit by hand.
// Source: https://raw.githubusercontent.com/yuhonas/free-exercise-db/main/dist/exercises.json (Unlicense / public domain)
// 873 exercises.

export const MUSCLES = [
  "neck",
  "traps",
  "shoulders",
  "chest",
  "biceps",
  "triceps",
  "forearms",
  "lats",
  "middle back",
  "lower back",
  "abdominals",
  "glutes",
  "quadriceps",
  "hamstrings",
  "adductors",
  "abductors",
  "calves"
] as const;
export const CATEGORIES = [
  "strength",
  "powerlifting",
  "olympic weightlifting",
  "strongman",
  "plyometrics",
  "stretching",
  "cardio"
] as const;
export const LEVELS = [
  "beginner",
  "intermediate",
  "expert"
] as const;
export const FORCES = [
  "push",
  "pull",
  "static"
] as const;
export const MECHANICS = [
  "compound",
  "isolation"
] as const;
export const EQUIPMENT = [
  "bands",
  "barbell",
  "body only",
  "cable",
  "dumbbell",
  "e-z curl bar",
  "exercise ball",
  "foam roll",
  "kettlebells",
  "machine",
  "medicine ball",
  "other"
] as const;

export type Muscle = (typeof MUSCLES)[number];
export type Category = (typeof CATEGORIES)[number];
export type Level = (typeof LEVELS)[number];
export type Force = (typeof FORCES)[number];
export type Mechanic = (typeof MECHANICS)[number];
export type Equipment = (typeof EQUIPMENT)[number];